Remove second compiler
