+++ demos/my_system/mix.exs
@@ -7,7 +7,7 @@ defmodule MySystem.MixProject do
       version: "0.1.0",
       elixir: "~> 1.10.0-dev",
       elixirc_paths: elixirc_paths(Mix.env()),
-      compilers: [:boundary_xref, :phoenix, :gettext] ++ Mix.compilers() ++ [:boundary],
+      compilers: [:boundary, :phoenix, :gettext] ++ Mix.compilers(),
       start_permanent: Mix.env() == :prod,
       deps: deps()
     ]
