+++ lib/boundary/mix_compiler.ex
@@ -1,5 +1,6 @@
 defmodule Boundary.MixCompiler do
   @moduledoc false
+
   # credo:disable-for-this-file Credo.Check.Readability.Specs
 
   def check(opts \\ []) do
