+++ lib/mix/tasks/compile/boundary.ex
@@ -21,7 +21,7 @@ defmodule Mix.Tasks.Compile.Boundary do
 
     def project do
       [
-        compilers: Mix.compilers() ++ [:boundary],
+        compilers: [:boundary] ++ Mix.compilers(),
         # ...
       ]
     end
@@ -39,7 +39,7 @@ defmodule Mix.Tasks.Compile.Boundary do
 
     def project do
       [
-        compilers: Mix.compilers() ++ extra_compilers(Mix.env()),
+        compilers: extra_compilers(Mix.env()) ++ Mix.compilers(),
         # ...
       ]
     end
@@ -86,10 +86,52 @@ defmodule Mix.Tasks.Compile.Boundary do
 
   @impl Mix.Task.Compiler
   def run(argv) do
-    errors = Boundary.MixCompiler.check()
-    print_diagnostic_errors(errors)
-    {status(errors, argv), errors}
+    BoundaryXref.start_link(path())
+    Mix.Task.Compiler.after_compiler(:app, &after_compiler(&1, argv))
+
+    tracers = Code.get_compiler_option(:tracers)
+    Code.put_compiler_option(:tracers, [__MODULE__ | tracers])
+
+    {:ok, []}
   end
+
+  @doc false
+  def trace({remote, meta, callee_module, name, arity}, env) when remote in ~w/remote_function remote_macro/a do
+    if env.module != nil do
+      BoundaryXref.add_call(
+        env.module,
+        %{callee: {callee_module, name, arity}, file: Path.relative_to_cwd(env.file), line: meta[:line]}
+      )
+    end
+
+    :ok
+  end
+
+  def trace(_event, _env), do: :ok
+
+  defp after_compiler({:ok, diagnostics}, argv) do
+    tracers = Enum.reject(Code.get_compiler_option(:tracers), &(&1 == __MODULE__))
+    Code.put_compiler_option(:tracers, tracers)
+
+    app = Keyword.fetch!(Mix.Project.config(), :app)
+    Application.load(app)
+    app_modules = MapSet.new(Application.spec(app, :modules))
+    BoundaryXref.finalize(app_modules)
+
+    calls =
+      BoundaryXref.calls(path())
+      |> Stream.map(fn {caller, meta} -> Map.put(meta, :caller_module, caller) end)
+      |> Stream.map(fn %{callee: {mod, _fun, _arg}} = entry -> Map.put(entry, :callee_module, mod) end)
+      |> Stream.reject(&(&1.callee_module == &1.caller_module))
+      |> Enum.map(&normalize_line/1)
+
+    errors = Boundary.MixCompiler.check(calls: calls)
+    print_diagnostic_errors(errors)
+    {status(errors, argv), diagnostics ++ errors}
+  end
+
+  defp normalize_line(%{line: {file, line}} = call), do: %{call | file: file, line: line}
+  defp normalize_line(call), do: call
 
   defp status([], _), do: :ok
   defp status([_ | _], argv), do: if(warnings_as_errors?(argv), do: :error, else: :ok)
@@ -120,4 +162,6 @@ defmodule Mix.Tasks.Compile.Boundary do
   defp severity(severity), do: [:bright, color(severity), "#{severity}: ", :reset]
   defp color(:error), do: :red
   defp color(:warning), do: :yellow
+
+  defp path(), do: Path.join(Mix.Project.build_path(), "boundary_calls.dets")
 end
